+++ src/eval-frame/tools/fetch-file-from-parent-context.js
@@ -1,0 +1,35 @@
+import { postMessageToEditor } from '../port-to-editor'
+
+const FETCH_RESOLVERS = {}
+
+function addResolvers(path, resolve, reject) {
+  FETCH_RESOLVERS[path] = { resolve, reject }
+}
+
+function getResolvers(path) {
+  return FETCH_RESOLVERS[path]
+}
+
+function deleteResolvers(path) {
+  delete FETCH_RESOLVERS[path]
+}
+
+export function onParentContextFileFetchSuccess(file, path) {
+  getResolvers(path).resolve(file)
+  deleteResolvers(path)
+}
+
+export function onParentContextFileFetchError(reason, path) {
+  getResolvers(path).reject(new Error(reason))
+  deleteResolvers(path)
+}
+
+export default async function fetchFileFromParentContext(path, fetchType) {
+  // used in eval frame.
+  return new Promise((resolve, reject) => {
+    // resolve and reject are handled in port-to-editor.js when
+    // the file is received by the editor.
+    addResolvers(path, resolve, reject)
+    postMessageToEditor('REQUEST_FETCH', { path, fetchType })
+  })
+}
